BBSecurinets/2023/bbjwt/requirements.txt
Flask==2.0.2
PyJWT==2.1.0
